# testing-service/docs/swagger.yaml
definitions:
...
  ErrorMessage:
    properties:
      errorDate:
        description: Time of the error, formatted as yyyy-mm-dd hh:mm:ss.mmm.
        example: "2024-05-17 11:33:23.038"
        type: string
      errorMessage:
        description: The error itself, which may be a custom message or an internal
          failure.
        example: Something went wrong.
        type: string
      serviceName:
        description: Name of the service that produced the message.
        example: testing-service
        type: string
      stacktrace:
        description: The stack trace, when one is available; otherwise "No Stacktrace
          Available".
        example: No Stacktrace Available
        type: string
    type: object
  Filter:
    properties:
      condition:
        type: string
      feature:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
  Header:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  ImportResult:
    properties:
      archive:
        type: string
      entityId:
        type: string
      entityName:
        type: string
      fileName:
        type: string
      message:
        type: string
      result:
        type: string
    type: object
  Matcher:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      entityName:
        type: string
      entityType:
        type: string
      id:
        type: string
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/MatcherParameter'
        type: array
      type:
        type: string
    type: object
  MatcherParameter:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  Message:
    properties:
      body:
        type: string
      headers:
        items:
          $ref: '#/definitions/Header'
        type: array
    type: object
  PathParameter:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  QueryParameter:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  RequestSettings:
    properties:
      message:
        $ref: '#/definitions/Message'
      method:
        type: string
      pathParameters:
        items:
          $ref: '#/definitions/PathParameter'
        type: array
      queryParameters:
        items:
          $ref: '#/definitions/QueryParameter'
        type: array
      timeout:
        type: integer
    type: object
...
  SelectionSpecification:
    properties:
      filters:
        items:
          $ref: '#/definitions/Filter'
        type: array
      ids:
        items:
          type: string
        type: array
      searchText:
        type: string
    type: object
...
  TestCase:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      requestSettings:
        $ref: '#/definitions/RequestSettings'
      responseValidationRules:
        items:
          $ref: '#/definitions/Matcher'
        type: array
      triggerReference:
        $ref: '#/definitions/TriggerReference'
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  TestCaseView:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      enabledRuleCount:
        type: integer
      id:
        type: string
      name:
        type: string
      requestSettings:
        $ref: '#/definitions/RequestSettings'
      responseValidationRules:
        items:
          $ref: '#/definitions/Matcher'
        type: array
      triggerReference:
        $ref: '#/definitions/TriggerReference'
      updatedAt:
        type: string
      updatedBy:
        type: string
      validationRuleCount:
        type: integer
    type: object
...
  TriggerReference:
    properties:
      chainId:
        type: string
      elementId:
        type: string
    type: object
...
info:
  contact: {}
  description: API of the testing service of the Qubership Integration Platform.
  title: Testing Service API
paths:
...
  /api/v1/test-cases:
    delete:
      consumes:
      - application/json
      operationId: deleteTestsCasesV1
      parameters:
      - description: Test cases IDs to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Delete test cases
      tags:
      - V1
      - Test Cases
    post:
      consumes:
      - application/json
      operationId: findTestsCasesV1
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort field
        enum:
        - id
        - name
        - description
        - enabled
        - chain_id
        - element_id
        - created_by
        - created_at
        - updated_by
        - updated_at
        - validation_rule_count
        - enabled_rule_count
        in: query
        name: sort_by
        type: string
      - default: ASC
        description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: sort_order
        type: string
      - description: Return IDs list
        in: query
        name: return_ids
        type: boolean
      - description: Test cases specification
        in: body
        name: specification
        required: true
        schema:
          $ref: '#/definitions/SelectionSpecification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TestCaseView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get test cases
      tags:
      - V1
      - Test Cases
  /api/v1/test-cases/{id}:
    delete:
      operationId: deleteTestCaseV1
      parameters:
      - description: Test case UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Delete test case
      tags:
      - V1
      - Test Cases
    get:
      operationId: findTestCaseByIdV1
      parameters:
      - description: Test case UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TestCaseView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get test case
      tags:
      - V1
      - Test Cases
    post:
      consumes:
      - application/json
      operationId: updateTestCaseV1
      parameters:
      - description: Test case UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Test case
        in: body
        name: testCase
        required: true
        schema:
          $ref: '#/definitions/TestCase'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TestCase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Update test case
      tags:
      - V1
      - Test Cases
  /api/v1/test-cases/create:
    post:
      consumes:
      - application/json
      operationId: createTestCaseV1
      parameters:
      - description: Test case
        in: body
        name: testCase
        required: true
        schema:
          $ref: '#/definitions/TestCase'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/TestCase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Create test case
      tags:
      - V1
      - Test Cases
  /api/v1/test-cases/export:
    post:
      consumes:
      - application/json
      operationId: exportTestCasesV1
      parameters:
      - description: Test cases IDs to export
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Export test cases
      tags:
      - V1
      - Test Cases
  /api/v1/test-cases/import:
    post:
      consumes:
      - multipart/form-data
      operationId: importTestCasesV1
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ImportResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Import test cases
      tags:
      - V1
      - Test Cases
...
produces:
- application/json
swagger: "2.0"
...